+++ src/iplugin.ts
@@ -1,7 +1,7 @@
 import type { URL } from 'node:url';
 import Summary from './summary.js';
 
-export interface IPlugin {
+export interface SummalyPlugin {
 	test: (url: URL) => boolean;
 	summarize: (url: URL, lang?: string) => Promise<Summary | null>;
 }