+++ src/plugins/index.ts
@@ -1,10 +1,10 @@
-import { IPlugin } from '@/iplugin.js';
+import { SummalyPlugin } from '@/iplugin.js';
 import * as amazon from './amazon.js';
 import * as wikipedia from './wikipedia.js';
 import * as branchIoDeeplinks from './branchio-deeplinks.js';
 
-export const plugins: IPlugin[] = [
-    amazon,
-    wikipedia,
-    branchIoDeeplinks,
+export const plugins: SummalyPlugin[] = [
+  amazon,
+  wikipedia,
+  branchIoDeeplinks,
 ];