+++ src/plugins/wikipedia.ts
@@ -25,7 +25,7 @@ export async function summarize(url: URL): Promise<summary> {
 	log(body);
 
 	if (!('query' in body) || !('pages' in body.query)) {
-		throw 'fetch failed';
+		throw new Error('fetch failed');
 	}
 
 	const info = body.query.pages[Object.keys(body.query.pages)[0]];