+++ src/summary.ts
@@ -1,8 +1,8 @@
 type Summary = {
 	/**
-	 * The description of that web page
+	 * The title of that web page
 	 */
-	description: string | null;
+	title: string | null;
 
 	/**
 	 * The url of the icon of that web page
@@ -10,9 +10,9 @@ type Summary = {
 	icon: string | null;
 
 	/**
-	 * The name of site of that web page
+	 * The description of that web page
 	 */
-	sitename: string | null;
+	description: string | null;
 
 	/**
 	 * The url of the thumbnail of that web page
@@ -20,14 +20,14 @@ type Summary = {
 	thumbnail: string | null;
 
 	/**
+	 * The name of site of that web page
+	 */
+	sitename: string | null;
+
+	/**
 	 * The player of that web page
 	 */
 	player: Player;
-
-	/**
-	 * The title of that web page
-	 */
-	title: string | null;
 
 	/**
 	 * Possibly sensitive
@@ -38,6 +38,13 @@ type Summary = {
 	 * The url of the ActivityPub representation of that web page
 	 */
 	activityPub: string | null;
+};
+
+export type SummalyResult = Summary & {
+	/**
+	 * The actual url of that web page
+	 */
+	url: string;
 };
 
 export default Summary;