+++ test/index.ts
@@ -46,6 +46,65 @@ afterEach(async () => {
 
 /* tests below */
 
+test('basic', async () => {
+	app = fastify();
+	app.get('/', (request, reply) => {
+		return reply.send(fs.createReadStream(_dirname + '/htmls/basic.html'));
+	});
+	await app.listen({ port });
+	expect(await summaly(host)).toEqual({
+		title: 'KISS principle',
+		icon: null,
+		description: null,
+		thumbnail: null,
+		player: {
+			url: null,
+			width: null,
+			height: null,
+			"allow": [
+			  "autoplay",
+			  "encrypted-media",
+			  "fullscreen",
+			],
+		},
+		sitename: 'localhost:3060',
+		sensitive: false,
+		url: host,
+		activityPub: null,
+	});
+});
+
+test('Stage Bye Stage', async () => {
+	// If this test fails, you must rewrite the result data and the example in README.md.
+
+	const summary = await summaly('https://www.youtube.com/watch?v=NMIEAhH_fTU');
+	expect(summary).toEqual(
+		{
+			"title": "【アイドルマスター】「Stage Bye Stage」(歌：島村卯月、渋谷凛、本田未央)",
+			"icon": "https://www.youtube.com/s/desktop/28b0985e/img/favicon.ico",
+			"description": "Website▶https://columbia.jp/idolmaster/Playlist▶https://www.youtube.com/playlist?list=PL83A2998CF3BBC86D2018年7月18日発売予定THE IDOLM@STER CINDERELLA GIRLS CG STAR...",
+			"thumbnail": "https://i.ytimg.com/vi/NMIEAhH_fTU/maxresdefault.jpg",
+			"player": {
+				"url": "https://www.youtube.com/embed/NMIEAhH_fTU?feature=oembed",
+				"width": 200,
+				"height": 113,
+				"allow": [
+					"autoplay",
+					"clipboard-write",
+					"encrypted-media",
+					"picture-in-picture",
+					"web-share",
+					"fullscreen",
+				]
+			},
+			"sitename": "YouTube",
+			"sensitive": false,
+			"activityPub": null,
+			"url": "https://www.youtube.com/watch?v=NMIEAhH_fTU"
+		}
+	);
+});
+
 test('faviconがHTML上で指定されていないが、ルートに存在する場合、正しく設定される', async () => {
 	app = fastify();
 	app.get('/', (request, reply) => {
@@ -391,5 +450,25 @@ describe('ActivityPub', () => {
 
 		const summary = await summaly(host);
 		expect(summary.activityPub).toBe(null);
+	});
+});
+
+describe('sensitive', () => {
+	test('default', async () => {
+		app = fastify();
+		app.get('/', (request, reply) => {
+			return reply.send(fs.createReadStream(_dirname + '/htmls/basic.html'));
+		});
+		await app.listen({ port });
+		expect((await summaly(host)).sensitive).toBe(false);
+	});
+
+	test('mixi:content-rating 1', async () => {
+		app = fastify();
+		app.get('/', (request, reply) => {
+			return reply.send(fs.createReadStream(_dirname + '/htmls/mixi-sensitive.html'));
+		});
+		await app.listen({ port });
+		expect((await summaly(host)).sensitive).toBe(true);
 	});
 });